feat(reading): support multithreading in reading scan functions to improve performance (#43)
Implemented multithreading for reading functions (read_vertices and read_edges) using DuckDB's built-in mechanisms (init_global & init_local). This allows chunks to be processed in parallel, with each chunk being read by a single thread at a time. As a result, performance on graphs with a large number of chunks now reaches and even exceeds that of raw DuckDB.